pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ap-copy-master-to-blink"
version = "0.1.0"
description = "Copy master calibration frames from library to blink directories"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]
dependencies = [
    "ap-common @ git+https://github.com/jewzaam/ap-common.git",
    "PyYAML>=6.0",
    "questionary>=2.0",
]

[project.scripts]
ap-copy-master-to-blink = "ap_copy_master_to_blink.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.11.2",
    "types-PyYAML>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ap_copy_master_to_blink*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313", "py314"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E266", "E501", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = false
warn_return_any = false
warn_unused_ignores = false
no_implicit_optional = false
